FIX: Correcciones crud unidad_medida

+++ backend/src/services/unidad_medida.service.js
@@ -6,9 +6,22 @@ export async function createUnidadMedidaService(data) {
     const unidadMedidaRepository = AppDataSource.getRepository(UnidadMedida);
 
     try {
-        const { nombre_unidad_medida } = data;
-        
+    
+         const { nombre_unidad_medida } = data;
+    /*Normalizar y convertir a minúsculas el nombre de la unidad de medida
+    const normalizedNombre = nombre_unidad_medida
+        .normalize("NFD")
+      .replace(/[\u0300-\u036f]/g, '')
+      .toLowerCase();
 
+    // Verificar si la unidad de medida ya existe
+    const existingUnidadMedida = await unidadMedidaRepository.findOneBy({
+      nombre_unidad_medida: normalizedNombre,
+    });
+
+    if (existingUnidadMedida) {
+      return [null, `La unidad de medida '${normalizedNombre}' ya existe.`];
+    } */
         const newUnidadMedida = unidadMedidaRepository.create({
             nombre_unidad_medida,
         });
@@ -25,15 +38,15 @@ export async function createUnidadMedidaService(data) {
 // Servicio para obtener todas las unidades de medida
 export async function getUnidadesMedidasService() {
     const unidadMedidaRepository = AppDataSource.getRepository(UnidadMedida);
-
+  
     try {
-        const unidades = await unidadMedidaRepository.find();
-        return [unidades, null];
+      const unidades = await unidadMedidaRepository.find();
+      return [unidades, null];
     } catch (error) {
-        console.error("Error al obtener las unidades de medida:", error);
-        return [null, error.message];
+      console.error("Error al obtener las unidades de medida:", error);
+      return [null, error.message];
     }
-}
+  }
 
 // Servicio para obtener una unidad de medida por ID
 export async function getUnidadMedidaByIdService(id) {
@@ -66,8 +79,13 @@ export async function deleteUnidadMedidaService(id) {
     const unidadMedidaRepository = AppDataSource.getRepository(UnidadMedida);
 
     try {
+        const unidadMedida = await unidadMedidaRepository.findOneBy({ id_unidad_medida: id });
+        if (!unidadMedida) {
+            return [null, `La unidad de medida con ID ${id} no existe.`];
+        }
+
         await unidadMedidaRepository.delete(id);
-        return [true, null];
+        return [unidadMedida, null];
     } catch (error) {
         console.error("Error al eliminar la unidad de medida:", error);
         return [null, error.message];
